refactor: 백엔드 서버 구조와 공통 입력 규칙 정리
- HTTP API 서버 구현을 라이브러리 server 모듈로 이동
- 입력 지원 확장자와 변환 옵션 검증 규칙을 공통화
- 단일 변환 내부 단계를 헬퍼로 분리해 중복을 줄임
- 백엔드 구조 변경 내용을 architecture와 memory 문서에 반영
