+++ docs/memory.md
@@ -12,6 +12,18 @@
 
 ## 최근 작업 로그
 
+### 2026-05-05 — 백엔드 서버/변환 코어 구조 리팩토링
+
+- 백엔드 Rust 코드만 대상으로 기능 변경 없는 구조 리팩토링 진행
+- 기존 `src/bin/server.rs` 의 HTTP API 구현을 `src/server.rs` 라이브러리 모듈로 이동
+  - `src/bin/server.rs` 는 `image_converter::server::run()` 만 호출하는 얇은 바이너리 엔트리포인트로 정리
+  - 서버 테스트는 서버 모듈 안에 유지해 라우터와 변환 API 계약을 계속 검증
+- 입력 지원 확장자 목록을 `src/input.rs` 로 공통화
+  - 배치 변환과 HTTP API 서버가 같은 지원 확장자 기준을 사용
+  - 사용자 메시지용 확장자 라벨도 공통 헬퍼로 제공
+- `JpegBackground::from_name_or_hex`, `ConversionOptions::for_format` 를 추가해 CLI/API 의 JPEG 배경색 검증 중복을 줄임
+- 단일 변환의 출력 검증, 이미지 로딩, 리사이즈, 인코딩/저장, 통계 생성 단계를 내부 헬퍼로 분리해 조용한 변환과 진행률 변환의 중복을 줄임
+
 ### 2026-05-05 — 웹 변환 작업대 프론트 구조 리팩토링
 
 - 프론트엔드만 대상으로 기능 변경 없는 구조 리팩토링 진행
