+++ src/lib.rs
@@ -7,6 +7,7 @@ pub mod error;
 pub mod format;
 pub mod input;
 pub mod interactive;
+pub mod server;
 pub mod utils;
 
 pub use batch::{
